LJ_GC64: Fix ir_khash for non-string GCobj.
Contributed by Peter Cawley.

(cherry picked from commit b4ed3219a1a98dd9fe7d1e3eeea3b82f5a780948)

When emitting the `IR_HREF` for a constant value lookup the `ir_khash()`
function is used to calculate the hash for the corresponding object.
This calculation must be the same as in the corresponding `hashkey()`
function from <lj_tab.c>.

Hash is calculated by passing two arguments `lo`, and `hi` to the
`hashrot()` routine. For non-string GC objects the first `lo` argument
is the same for GC64 and not GC64 mode -- lower 32 bits of the object
address. For GC64 mode `hi` argument is upper 32 bits of the object
address, including a specific type NaN-tag. This `hi` argument in
`ir_khash()` function is miscalculated in GC64 mode using a non-GC64
value (`lo` + `HASH_BIAS`). As a result, the hash for the GC object is
miscalculated on trace and we exit from the trace due to an assertion
guard on the type or value check.

This patch fixes calculation of the hash value on trace for GC64 mode by
making it consistent with the `hashkey()`.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#7230

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 19f0914670e700c997aede0a66b7bf1699a673e2)