ARM64: Fix write barrier in BC_USETS.
Contributed by Javier Guerra Giraldez.

(cherry picked from commit c785131ca5a6d24adc519e5e0bf1b69b671d912f)

Closed upvalues are never gray. Hence, when closed upvalue is marked, it
is marked as black. Black objects can't refer white objects, so for
storing a white value in a closed upvalue, we need to move the barrier
forward and color our value to gray by using `lj_gc_barrieruv()`. This
function can't be called on closed upvalues with non-white values since
there is no need to mark it again.

USETS bytecode for arm64 architecture has the incorrect NZCV condition
flag value in the instruction that checks the upvalue is closed:
| tst TMP1w, #LJ_GC_WHITES
| ccmp TMP0w, #0, #0, ne
| beq <1 // branch out from barrier movement
`TMP0w` contains `upvalue->closed` field, so the upvalue is open if this
field equals to zero (the first one in `ccmp`). The second zero is the
value of NZCV condition flags[1] yielded if the specified condition
(`ne`) is met for the current values of the condition flags[2]. Hence,
if the value to be stored is not white (`TMP1w` holds its color), then
the condition is FALSE and all flags bits are set to zero so the branch
is not taken (Zero flag is not set). If this happens at propagate or
atomic GC phase, the `lj_gc_barrieruv()` function is called and the gray
value to be set is marked like if it is white. That leads to the
assertion failure in the `gc_mark()` function.

This patch changes NZCV condition flag to 4 (Zero flag is set) to take
the correct branch after `ccmp` instruction.

Sergey Kaplun:
* added the description and the test for the problem

[1]: https://community.arm.com/developer/ip-products/processors/b/processors-ip-blog/posts/condition-codes-1-condition-flags-and-codes
[2]: https://developer.arm.com/documentation/dui0801/g/pge1427897656225

Part of tarantool/tarantool#5629

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>