FFI/ARM64: Fix pass-by-value struct calling conventions.
(cherry picked from commit 9143e86498436892cb4316550be4d45b68a61224)

If the argument type is a Composite Type then the size of the
argument is rounded up to the nearest multiple of 8 bytes. LuaJIT FFI
backend makes this rounding unconditionally for arm64 architecture and
uses the result value to determine the necessary amount of registers
for the call.

The arm64 parameters passing rules for Homogeneous Floating-point
Aggregates (HFA) are the following [1]:
| If the argument is an HFA and there are sufficient unallocated
| Floating-point registers, then the argument is allocated to
| Floating-point registers (with one register per member of the HFA).

So for the HFA composed of 3 32-bit float members the extra one (4th)
register is supposed as occupied.

When the second parameter is passed to the function, these fields are
loaded starting from 5th register. As far as the procedure to be called
uses 6 registers according to the ABI it leads to the wrong value of the
second parameter passed to the callee (0 due to the corresponding memset
in `cconv_struct_init()`).

This patch fixes this case by using the real size of structure in the
calculation of necessary amount of registers to use.

Sergey Kaplun:
* added the description and the test for the problem

[1]: https://github.com/ARM-software/abi-aa/blob/main/aapcs64/aapcs64.rst#parameter-passing-rules

Part of tarantool/tarantool#6548

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 9b8e209042b9dacd21012399bcc626bd2ade07b1)