profile: introduce symtab write module
This patch adds profile writer that writes all necessary Lua
functions prototypes info like GCproto address, name of the chunk this
function was defined in and number of the first line of it.
See <ljp_symtab.h> for details.

Usage of this module will be added at the next patches.

Part of tarantool/tarantool#5442