tools: introduce a memory profile parser
This patch adds a parser for binary data dumped via the memory profiler.
It is a set of the following Lua modules:
* utils/bufread.lua: read binary data from a file.
* utils/symtab.lua: symbol table decode functions
* memprof/parse.lua: decode the memory profiler event stream
* memprof/humanize.lua: display decoded data in human readable format
* memprof.lua: Lua script to display data

There is also a stand-alone bash script <luajit-parse-memprof> that
displays human readable parsed data to a stdout. It calls <memprof.lua>
with a corresponding LUA_PATH.

Part of tarantool/tarantool#5442
Part of tarantool/tarantool#5490

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>