test: fix Lua command in utils.selfrun
Now <utils.selfrun> doesn't respect the flags passed to LuaJIT. If one
is given, the command to be executed is completely broken, since Lua
executable is moved to the slots less than -1. This patch fixes the
function behaviour and scans all slots in <arg> table that are less than
-1 until nil is met.

There is a minor flaw in the introduced helper: though the flag value
can be wrapped with the quotes when the script is run via shell, the
restored command doesn't have any, since they are removed before LuaJIT
is invoked. Such shell specific behaviour looks to be out of the scope
of this changeset.

Follows up tarantool/tarantool#4862

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>