test: adapt PUC-Rio test with activeline check
LuaJIT does not report the line with single "end" statement (the last
line of the function) as an active line in `debug.getinfo()`, unlike Lua
does. There is no bytecode related to this line, so it is "unreachable"
and may be considered not active.

This patch excludes the last line of the function from the check,
considering LuaJIT's behaviour.

Resolves tarantool/tarantool#5708
Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 67c41b788c9daec225cdaef8fed6f3730626113b)