test: disable PUC-Rio test for fast function name
When LuaJIT tries to detect function's name for error message, it
determines the bytecode position first, via `debug_framepc()`. For
tailcall there is no creation of additional Lua frame, so the current
frame is a C protected frame (at the moment of the error throwing).
`debug_framepc()` can't determine bytecode position for non-Lua frames,
so the fast function name is not detected.

This patch disables the test that checks the name of built-in functions
reported in error in tailcall.

Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 31cb65dadc1e7f417fc5418bcfcb9c83eeff1cd1)