test: refactor with _retrieve_progname
The patch[1] from the lua-Harness suite fixes the same issue as commit
837688590919fcf8de47ef90479fd2a640c8fddc ('test: adjust lua-Harness
suite to CMake machinery') does except the single difference:
`get_lua_binary_name()` is called `_retrieved_progname()`. As a result
of this patch the function is renamed to follow the original naming.

[1]: https://framagit.org/fperrad/lua-Harness/-/commit/1be25a8

Part of tarantool/tarantool#5970
Part of tarantool/tarantool#4473

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>