test: support tarantool cli in lua-Harness
The patch[1] from the lua-Harness suite adjusts some checks in
the lua-Harness tests, so they are compatible with Tarantool now.

As a result, the introduced assertion for Tarantool error message fails
on MacOS, since getopt_long(3) yields the error message without single
quotes wrapping the flag. Hence, this commit also includes the fix for
that from the patch[2] in the trunk.

[1]: https://framagit.org/fperrad/lua-Harness/-/commit/1da5b1b
[2]: https://framagit.org/fperrad/lua-Harness/-/commit/a0532c5

Part of tarantool/tarantool#5970
Part of tarantool/tarantool#4473

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>