tools: add cli flag to run profile dump parsers
It is really inconvenient to use a standalone shell script to parse
binary dumps from profilers. That is why this commit introduces a
CLI flag for tools in the LuaJIT, so now it is possible to parse
a memprof dump as simple as:
```
luajit -tm memprof.bin
luajit -tm --leak-only memprof.bin
```

And the sysprof too:
```
luajit -ts sysprof.bin
```

The scripts `luajit-parse-memprof` and `luajit-parse-sysprof`
are purged as a result of these changes.

Resolves tarantool/tarantool#5688

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 0380a8c17b93f4e839838c03817e740ef626d0ce)