gdb: support LJ_DUALNUM mode
luajit-gdb.py displays integers in LJ_DUALNUM mode as nan-s. The dumper
function produces output considering any input value as a double.
However, in LJ_DUALNUM mode, integers and doubles are stored
differently, so the `itype` of a double must be less than `LJ_TISNUM`,
and the `itype` of an integer must be `LJ_TISNUM`. With this fact in
mind, we can easily differentiate one from another.

Closes tarantool/tarantool#6224

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>