tools: introduce parsers for new sampling profiler
Since the sysprof's binary output is not human-readable, so there is a
demand to create a parser. The parser, which this commit provides,
converts sysprof's event stream into the format that the flamegraph.pl
can process.

The sysprof parser machinery uses the same symtab module as memprof's
parser since the format is the same.

Resolves of tarantool/tarantool#781

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>