test: fix path storage for non-concatable objects
When a key of the table to be compared via `tap.test:is_deeply()` is a
non-concatable object (i.e. lightuserdata) its concatenation with the
path prefix raises an error. As a result of this patch the object is
explicitly converted to string to avoid this error.

Needed for tarantool/tarantool#2712

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 9ddf35689c6ffbc243eeef79b112b78ebcbc648a)