vm: save topframe info into global_State
Since commit 111d377d524e54e02187148a1832683291d620b2 ('vm: introduce VM
states for Lua and fast functions') the VM has LFUNC and FFUNC states.

The upcoming sampling profiler uses these vmstates to determine if the
guest stack is valid and can be dumped or not. But even if it meets
vmstates considered "valid", it needs a consistent stack base pointer to
dump the Lua calling stack: the cur_L->base is synced only when the
control is passed from the VM to users code, but the sampling signal may
arrive at any moment of time, so cur_L->base may be inconsistent.

This patch extends the global_State structure with a field `top_frame`,
which stores a valid stack base pointer while VM is in LFUNC, CFUNC or
FFUNC states. For this reason, a new macros `set_vmstate_sync_base`
added to vm_x64 and vm_x86 (other architectures are NYI at the moment)
to sync the BASE register before setting the VM state.

Part of tarantool/tarantool#781

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>