memprof: refactor location parsing
As it is hard to keep function `id_location` easily usable in the
function `parse_location` with a growing number of fields in the
<loc> table, the code is refactored to make it more understandable.

The API of <utils/symtab.lua> module changed with adding a public
function `id(loc)` returning the same id-string, as `id_location`
did. This function is needed to be used in other modules to
"stringify" the location and use it as a key in the key/value
storage.

Part of tarantool/tarantool#5814

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>