test: adapt PUC-Rio test with reversed function
The first fiber in Tarantool has only 512Kb of the stack, which is not
enough to handle such a deep call chain. Thus, the chunk of code being
considered cores with SIGSEGV as a result of C stack overflow. The test
is adapted to Tarantool by decreasing the string length.

Resolves tarantool/tarantool#5782
Part of tarantool/tarantool#5870

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 951733d7f226bee9760951cf0af1981001b650f4)