memprof: add symbol generations
There are situations, when a new prototype or trace has the same address
than the previously dumped one. In this case, the new symtab entry will
shadow the preceding one, and if we demangle symbols after all the
events are parsed, we might obtain irrelevant information.

To avoid such situation as a result of this change the symtab stores not
a single entry, but the array of entries for each location. The location
structure <loc> was extended by the field `gen` representing the
generation of symtab entry associated with this particular location
entry. Also, the <utils/symtab.lua> API has been extended by function
`symtab.loc(symbols, args)` so that it can create a new location entity
with the proper generation value.

Part of tarantool/tarantool#5815

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>