memprof: extend symtab with C symbols
This commit provides an ability to enrich memprof's symbol table with
information about C symbols. The parser is updated correspondingly.

If there is .symtab section or at least .dynsym segment in a shared
library, then memprof dumps its contents in symtab. The <sym> was
extended the following way:
| sym := sym-lua | sym-cfunc | sym-trace | sym-final
| sym-cfunc := sym-header sym-addr sym-name

If none of those are present, then instead of a symbol name and its
address there will be name and address of a shared library containing
that symbol.

Resolves tarantool/tarantool#5813

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>