symtab: fix .symtab section dump of the executable
The `dl_iterate_phdr` returns an empty string as a name for the
executable from which it was called. It is still possible to access its
dynamic symbol table, but it is vital for sysprof to obtain the main
symbol table for the LuaJIT executable. To do so, we need a valid path
to the executable.

Since there is no way to obtain the path to a running executable using
the C standard library, this commit adds call to `readlink` to gather
the symbolic link from `/proc/self/exe`. Most of the UNIX-based systems
have procfs, so it is not a problem.

Needed for tarantool/tarantool#7244

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 799ead84ac8dc2a76a83679d843898f11432be06)