test: fix tarantool suite for out of source build
jit/vmdef.lua is autogenerated file, so it's put to src/ directory
located in the scope of the binary artefacts tree. Before the patch
LUA_PATH lacks this path, so tarantool-tests target fails due to
jit/vmdef.lua misseek. As a result of this change src/ directory in the
scope of the binary tree is included to LUA_PATH as well as the one from
the source tree has been.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit c49832f2a6406f91f2b9620f9115b51de23b0a49)