lldb: introduce luajit-lldb
It is impossible to run gdb on M1 devices, the only available
debugger is lldb. The luajit-gdb extension doesn't work with
lldb, so this patch introduces the luajit-lldb extension,
which re-implements exactly the same functionality.

It is worth noting that the extension is called
`luajit_lldb.py` instead of `luajit-lldb.py` because lldb
imports it as a Python module, and it is prohibited to use
dashes in a python module name.

Part of tarantool/tarantool#4808

Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 62fc84a8f89b8e5650162ba1c7696b0f84cf5c25)