sysprof: replace `backtrace` with libunwind
`backtrace` fails to unwind the host stack in LuaJIT since there
are no frame pointers during the VM calls, even with the
`-fno-omit-frame-pointer` option set. Sometimes, those failures
cause crashes. Moreover, `backtrace` is not signal-safe, which is
vital for the sysprof. This commit replaces it with the
libunwind-based unwinder, which makes use of additional runtime
information to provide robust unwinding even if there are no frame
pointers and is signal-safe. Because Tarantool uses its own fiber
unwinder, the new default unwinder is not built into bundled
builds.

Also, this commit enables C API tests, which used to crash with
`backtrace`, and disables an assertion regarding the successful
backtracer function setting, as it is unnecessary.

Part of tarantool/tarantool#781