sysprof: fix interval parsing in dual-number mode
When a profiling interval is parsed, `tvisnum()` is used to check that
the given value is a number. But in case of dual-number mode this check
returns false as far as the given value is an integer, so `tvisnumber()`
should be used to cover both cases.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 6cb7f989b8fb3b58f7169aa88d1188623179412c)