test: introduce test:skipall TAP helper
Sometimes one needs to skip all tests in the script. The TAP13 protocol
supports this via the following line:
| 1..0 # SKIP <reason>

Within this commit <test:skipall> helper is introduced that yields the
aforementioned line to the TAP consumer and exits the test script. It's
worth mentioning, that skipall helper is implemented only for root (i.e.
test with no parent) tests and subtests are not supported at the moment.
The corresponding assertion with a FIXME reason is added to <finalize>.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 8d5a92b85808a043c269691a7ec40d4ef44d5b64)