test: introduce test:skiprest TAP helper
Sometimes one needs to skip the remaining tests in the script. There is
no special stub in the TAP13 protocol for this, but this could be done via
skipping all the remaining planned tests.

Within this commit <test:skiprest> helper is introduced that calls
<test:skip> (planned - total) times and exits the test script. As for
the previous commit it's worth mentioning, that skiprest helper is
implemented only for root (i.e. test with no parent) tests and subtests
are not supported at the moment. The corresponding assertion with FIXME
reason can be found in <finalize> routine.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 8a0b68696b1c2dcf2fd42adee68e85884ad8f765)