Print errors from __gc finalizers instead of rethrowing them.
Finalizers are not supposed to throw errors -- this is undefined behavior.
Lua 5.1 - 5.3 and (previously) LuaJIT rethrow the error. This randomly
breaks some unrelated code that just happens to do an allocation. Bad.
Lua 5.4 catches the error and emits a warning instead. But warnings are
not enabled by default, so it fails silently. Even worse.
LuaJIT (now) catches the error and emits a VM event. The default event
handler function prints "ERROR in finalizer: ...".
Set a custom handler function with: jit.attach(handler, "errfin")

(cherry-picked from commit 1c279127050e86e99970100e9c42e0f09cd54ab7)

The default handler for finalizer errors is set during the
Lua initialization. Namely, in the `luaL_newstate`.

Along with the introduction of the new `ERRFIN` VM event, the high
bits for the old VM events are removed since they are scratched
anyway by the bitwise operation `(hash)<<3` in the `VMEVENT_DEF`
macro.

This patch results in a regression in the PUC-Rio test suite. The
test in the suite for the error in the GC finalizer fails after
the patch because the error is now handled with the VM event
handler instead of being rethrown. Hence, the `collectgarbage`
finishes successfully despite the error in the GC finalizer.
Considering this change, the test was disabled.

There is also another regression in the `misclib-getmetrics-capi`,
because there are a few test cases reliant on the `lua_gettop(L)`
value, which is broken after this patch. The `_VMEVENTS` table,
where the error handler for GC finalizers is set, was not cleared
from the stack after the initialization. This issue is fixed in
the following patch.

Maxim Kokryashkin:
* added the test for the problem and the description for the patch

Part of tarantool/tarantool#9145

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 58d6bd051ae625ba06b576c5722254ca1ae0098e)