sysprof: fix crash during FFUNC stream
Sometimes, the Lua stack can be inconsistent during the FFUNC execution,
which may lead to a sysprof crash during the stack unwinding.

This patch replaces the `top_frame` property of `global_State` with
`lj_sysprof_topframe` structure, which contains `top_frame` and `ffid`
properties. `ffid` property makes sense only when the LuaJIT VM state is
set to `FFUNC`. That property is set to the ffid of the fast function
that VM is about to execute.  In the same time, `top_frame` property is
not updated now, so the top frame of the Lua stack can be streamed based
on the ffid, and the rest of the Lua stack can be streamed as usual.

Also, this patch fixes the build via Makefile.original by adding the
`LJ_HASSYSPROF` flag support to it.

Resolves tarantool/tarantool#8594

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 285a1b0a16c1c3224ab93f7cf1fef17ab29e4ab4)